Fix bug in resign logic
The bot lost against Brown (random player) as it resigned when all
simulated moves lead to a loss. At the end of the game this can however
happen when the only possible moves are plays in the opponents eyes.

So now we only resign if both all simulations lead to a loss and playing
a pass move would lead to a loss.