Don't store the game state in the tree
It's not necessary to store any game state (be it Game or Board) in the
tree. Collecting the moves to playing during tree descent is good enough
and much quicker. With the current implementation of the Board we do
however can't do super ko violation checks in the game tree.