Calculate 95% and 99% percent confidence intervals from test data
It turns out that the error that gogui-twogtp reports is just the
standard deviation. This of course means that the result will fall
within that range 68% of the time only!

I've now enhanced my calculation script to report both the 2sigma (95%
confidence) and 3sigma (99% confidence) intervals.

It appears that just running 200 games isn't really good
enough (anymore).