Remove plays again
plays was very similar to playouts. It just included a scaling
factor to nudge the bot towards trying to produce wins that have
a larger margin. That obviously didn't work so I think it makes
sense to throw it out as it's just added complexity for no gain.

I also removed the `Node#mark_as_terminal` function because I
couldn't figure out what it's purpose was. We'll have to wait for
the benchmark results to see if I messed up. ;)